@@ -4,15 +4,14 @@ import 'package:rxdart/rxdart.dart';
 import 'FormControl.dart';
 import 'Validator.dart';
 
-
 class FormGroup {
-  Map<String, FormControl> _controls = new Map<String, FormControl>();
-  Future Function(dynamic data) _submitCallback;
+  Map<String?, FormControl> _controls = new Map<String?, FormControl>();
+  Future Function(dynamic data)? _submitCallback;
   // Stream<bool> get dirtyStream => _dirty.stream;
-  StreamSubscription _submitEnabledSubscription;
-  StreamSubscription _controlsStatusSubscription;
+  StreamSubscription? _submitEnabledSubscription;
+  StreamSubscription? _controlsStatusSubscription;
   final _statusStream = BehaviorSubject<ControlStatus>.seeded(ControlStatus.valid);
-  Validator _validator;
+  Validator? _validator;
   // FutureValidator _futureValidator;
 
   Stream<ControlStatus> get statusStream => _statusStream;
@@ -21,9 +20,9 @@ class FormGroup {
   bool get invalid => _statusStream.value == ControlStatus.invalid;
   bool get pending => _statusStream.value == ControlStatus.pending;
 
-  final _errorStream = BehaviorSubject<String>.seeded('');
-  Stream<String> get error => _errorStream;
-  String getError() => _errorStream.value;
+  final _errorStream = BehaviorSubject<String?>.seeded('');
+  Stream<String?> get error => _errorStream;
+  String? getError() => _errorStream.value;
 
   final _submitting = BehaviorSubject<bool>.seeded(false);
   Stream<bool> get submittingStream => _submitting;
@@ -35,16 +34,15 @@ class FormGroup {
 
   FormGroup();
 
-
   void add(FormControl control) {
-    _controls.addEntries([new MapEntry<String, FormControl>(control.name, control)]);
+    _controls.addEntries([new MapEntry<String?, FormControl>(control.name, control)]);
     _initStatusStream();
     validate();
   }
 
   void addAll(List<FormControl> controls) {
     controls.forEach((control) {
-      _controls.addEntries([new MapEntry<String, FormControl>(control.name, control)]);
+      _controls.addEntries([new MapEntry<String?, FormControl>(control.name, control)]);
     });
     _initStatusStream();
     validate();
@@ -73,41 +71,38 @@ class FormGroup {
       validate();
     }
   }
-  
+
   void _initStatusStream() {
     final streams = _controls.map((n, c) => MapEntry(n, c.statusStream)).values.toList();
     if (_controlsStatusSubscription != null) {
-      _controlsStatusSubscription.cancel();
+      _controlsStatusSubscription!.cancel();
       _controlsStatusSubscription = null;
     }
-    _controlsStatusSubscription = Observable
-      .merge(streams)
-      .listen((value) {
-        final old = _statusStream.value;
-        if (value != old) {
-          final hasInvalid = _controls.values.where((c) => c.invalid).length > 0;
-          if (hasInvalid) {
-            _statusStream.add(ControlStatus.invalid);
-            _runValidator();
-            return;
-          }
-
-          final hasPending = _controls.values.where((c) => c.pending).length > 0;
-          if (hasPending) {
-            _statusStream.add(ControlStatus.pending);
-          }
-          else {
-            _runValidator();
-            _setStatus();
-          }
+    _controlsStatusSubscription = Rx.merge(streams).listen((value) {
+      final old = _statusStream.value;
+      if (value != old) {
+        final hasInvalid = _controls.values.where((c) => c.invalid).length > 0;
+        if (hasInvalid) {
+          _statusStream.add(ControlStatus.invalid);
+          _runValidator();
+          return;
         }
-      });
+
+        final hasPending = _controls.values.where((c) => c.pending).length > 0;
+        if (hasPending) {
+          _statusStream.add(ControlStatus.pending);
+        } else {
+          _runValidator();
+          _setStatus();
+        }
+      }
+    });
 
     if (_submitEnabledSubscription != null) {
-      _submitEnabledSubscription.cancel();
+      _submitEnabledSubscription!.cancel();
       _submitEnabledSubscription = null;
     }
-    _submitEnabledSubscription = Observable.merge([_statusStream, _submitting]).listen((data) {
+    _submitEnabledSubscription = Rx.merge([_statusStream, _submitting]).listen((data) {
       final newValue = status == ControlStatus.valid && !submitting;
       if (_submitEnabledStream.value != newValue) {
         _submitEnabledStream.value = newValue;
@@ -116,15 +111,14 @@ class FormGroup {
   }
 
   bool _hasError() {
-    return _errorStream.value != null && _errorStream.value.isNotEmpty;
+    return _errorStream.value != null && _errorStream.value!.isNotEmpty;
   }
 
   void _runValidator() {
     final value = getValue();
     if (_validator != null) {
-      _errorStream.value = _validator.validate(value);
-    }
-    else if (_errorStream.value != null) {
+      _errorStream.value = _validator!.validate(value);
+    } else if (_errorStream.value != null) {
       _errorStream.value = null;
     }
   }
@@ -132,8 +126,7 @@ class FormGroup {
   void _setStatus() {
     if (_hasError()) {
       _statusStream.add(ControlStatus.invalid);
-    }
-    else {
+    } else {
       // todo: review
       // if (_futureValidator != null) {
       //   _statusStream.add(ControlStatus.pending);
@@ -148,8 +141,8 @@ class FormGroup {
       //   });
       // }
       // else {
-        _errorStream.value = '';
-        _statusStream.add(ControlStatus.valid);
+      _errorStream.value = '';
+      _statusStream.add(ControlStatus.valid);
       // }
     }
   }
@@ -162,21 +155,22 @@ class FormGroup {
     return _controls.containsKey(name);
   }
 
-  FormControl<T> get<T>(String name) {
+  FormControl<T>? get<T>(String name) {
     if (!_controls.containsKey(name)) {
-      throw Exception("FormControl with name: $name not found. don't forget to add it to the formGroup: formGroup.addControl(FormControl<Type>('$name', ''))");
+      throw Exception(
+          "FormControl with name: $name not found. don't forget to add it to the formGroup: formGroup.addControl(FormControl<Type>('$name', ''))");
     }
-    return _controls[name];
+    return _controls[name] as FormControl<T>?;
   }
 
   void setValue(Map<String, Object> value) {
     _controls.forEach((name, c) {
-      c.value = value[name];
+      c.value = value[name!];
     });
   }
 
-  Map<String, Object> getValue() {
-    Map<String, Object> values = new Map<String, Object>();
+  Map<String?, Object?> getValue() {
+    Map<String?, Object?> values = new Map<String?, Object?>();
     _controls.forEach((name, c) {
       values[name] = c.value;
     });
@@ -196,12 +190,12 @@ class FormGroup {
       return;
     }
     _submitting.value = true;
-    
+
     await validate();
     if (valid) {
-      await _submitCallback(getValue());
+      await _submitCallback!(getValue());
     }
-    
+
     _submitting.value = false;
   }
 
@@ -214,11 +208,11 @@ class FormGroup {
     _submitting.close();
     _errorStream.close();
     if (_controlsStatusSubscription != null) {
-      _controlsStatusSubscription.cancel();
+      _controlsStatusSubscription!.cancel();
       _controlsStatusSubscription = null;
     }
     if (_submitEnabledSubscription != null) {
-      _submitEnabledSubscription.cancel();
+      _submitEnabledSubscription!.cancel();
       _submitEnabledSubscription = null;
     }
     _controls.forEach((name, c) {
